+++ binary/bit_manipulation/README.md
